# docs/openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "JobFit RAG API",
                 "version":  "0.5.0"
             },
    "paths":  {
...
                  "/api/analyze":  {
                                       "post":  {
                                                    "summary":  "Analyze",
                                                    "operationId":  "analyze_api_analyze_post",
                                                    "requestBody":  {
                                                                        "content":  {
                                                                                        "application/json":  {
                                                                                                                 "schema":  {
                                                                                                                                "$ref":  "#/components/schemas/AnalyzeRequest"
                                                                                                                            }
                                                                                                             }
                                                                                    },
                                                                        "required":  true
                                                                    },
                                                    "responses":  {
                                                                      "200":  {
                                                                                  "description":  "Successful Response",
                                                                                  "content":  {
                                                                                                  "application/json":  {
                                                                                                                           "schema":  {
                                                                                                                                          "$ref":  "#/components/schemas/AnalysisReport-Output"
                                                                                                                                      }
                                                                                                                       }
                                                                                              }
                                                                              },
                                                                      "422":  {
                                                                                  "description":  "Validation Error",
                                                                                  "content":  {
                                                                                                  "application/json":  {
                                                                                                                           "schema":  {
                                                                                                                                          "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                      }
                                                                                                                       }
                                                                                              }
                                                                              }
                                                                  }
                                                }
                                   },
                  "/api/resume-matrix":  {
                                             "post":  {
                                                          "summary":  "Resume Matrix",
                                                          "operationId":  "resume_matrix_api_resume_matrix_post",
                                                          "requestBody":  {
                                                                              "content":  {
                                                                                              "application/json":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/ResumeMatrixRequest"
                                                                                                                                  }
                                                                                                                   }
                                                                                          },
                                                                              "required":  true
                                                                          },
                                                          "responses":  {
                                                                            "200":  {
                                                                                        "description":  "Successful Response",
                                                                                        "content":  {
                                                                                                        "application/json":  {
                                                                                                                                 "schema":  {
                                                                                                                                                "$ref":  "#/components/schemas/ResumeMatrixReport"
                                                                                                                                            }
                                                                                                                             }
                                                                                                    }
                                                                                    },
                                                                            "422":  {
                                                                                        "description":  "Validation Error",
                                                                                        "content":  {
                                                                                                        "application/json":  {
                                                                                                                                 "schema":  {
                                                                                                                                                "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                            }
                                                                                                                             }
                                                                                                    }
                                                                                    }
                                                                        }
                                                      }
                                         },
                  "/api/parse-resume":  {
                                            "post":  {
                                                         "summary":  "Parse Resume",
                                                         "operationId":  "parse_resume_api_parse_resume_post",
                                                         "requestBody":  {
                                                                             "content":  {
                                                                                             "multipart/form-data":  {
                                                                                                                         "schema":  {
                                                                                                                                        "$ref":  "#/components/schemas/Body_parse_resume_api_parse_resume_post"
                                                                                                                                    }
                                                                                                                     }
                                                                                         },
                                                                             "required":  true
                                                                         },
                                                         "responses":  {
                                                                           "200":  {
                                                                                       "description":  "Successful Response",
                                                                                       "content":  {
                                                                                                       "application/json":  {
                                                                                                                                "schema":  {
                                                                                                                                               "$ref":  "#/components/schemas/ParseResumeResponse"
                                                                                                                                           }
                                                                                                                            }
                                                                                                   }
                                                                                   },
                                                                           "422":  {
                                                                                       "description":  "Validation Error",
                                                                                       "content":  {
                                                                                                       "application/json":  {
                                                                                                                                "schema":  {
                                                                                                                                               "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                           }
                                                                                                                            }
                                                                                                   }
                                                                                   }
                                                                       }
                                                     }
                                        },
...
                  "/api/reports/{report_id}":  {
                                                   "get":  {
                                                               "summary":  "Report Detail",
                                                               "operationId":  "report_detail_api_reports__report_id__get",
                                                               "parameters":  [
                                                                                  {
                                                                                      "name":  "report_id",
                                                                                      "in":  "path",
                                                                                      "required":  true,
                                                                                      "schema":  {
                                                                                                     "type":  "string",
                                                                                                     "title":  "Report Id"
                                                                                                 }
                                                                                  }
                                                                              ],
                                                               "responses":  {
                                                                                 "200":  {
                                                                                             "description":  "Successful Response",
                                                                                             "content":  {
                                                                                                             "application/json":  {
                                                                                                                                      "schema":  {
                                                                                                                                                     "$ref":  "#/components/schemas/AnalysisReport-Output"
                                                                                                                                                 }
                                                                                                                                  }
                                                                                                         }
                                                                                         },
                                                                                 "422":  {
                                                                                             "description":  "Validation Error",
                                                                                             "content":  {
                                                                                                             "application/json":  {
                                                                                                                                      "schema":  {
                                                                                                                                                     "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                 }
                                                                                                                                  }
                                                                                                         }
                                                                                         }
                                                                             }
                                                           },
                                                   "delete":  {
                                                                  "summary":  "Remove Report",
                                                                  "operationId":  "remove_report_api_reports__report_id__delete",
                                                                  "parameters":  [
                                                                                     {
                                                                                         "name":  "report_id",
                                                                                         "in":  "path",
                                                                                         "required":  true,
                                                                                         "schema":  {
                                                                                                        "type":  "string",
                                                                                                        "title":  "Report Id"
                                                                                                    }
                                                                                     }
                                                                                 ],
                                                                  "responses":  {
                                                                                    "200":  {
                                                                                                "description":  "Successful Response",
                                                                                                "content":  {
                                                                                                                "application/json":  {
                                                                                                                                         "schema":  {
                                                                                                                                                        "type":  "object",
                                                                                                                                                        "additionalProperties":  {
                                                                                                                                                                                     "type":  "boolean"
                                                                                                                                                                                 },
                                                                                                                                                        "title":  "Response Remove Report Api Reports  Report Id  Delete"
                                                                                                                                                    }
                                                                                                                                     }
                                                                                                            }
                                                                                            },
                                                                                    "422":  {
                                                                                                "description":  "Validation Error",
                                                                                                "content":  {
                                                                                                                "application/json":  {
                                                                                                                                         "schema":  {
                                                                                                                                                        "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                    }
                                                                                                                                     }
                                                                                                            }
                                                                                            }
                                                                                }
                                                              }
                                               },
...
                  "/api/reports-import":  {
                                              "post":  {
                                                           "summary":  "Import Report Bundle",
                                                           "operationId":  "import_report_bundle_api_reports_import_post",
                                                           "requestBody":  {
                                                                               "content":  {
                                                                                               "application/json":  {
                                                                                                                        "schema":  {
                                                                                                                                       "$ref":  "#/components/schemas/ReportsExport-Input"
                                                                                                                                   }
                                                                                                                    }
                                                                                           },
                                                                               "required":  true
                                                                           },
                                                           "responses":  {
                                                                             "200":  {
                                                                                         "description":  "Successful Response",
                                                                                         "content":  {
                                                                                                         "application/json":  {
                                                                                                                                  "schema":  {
                                                                                                                                                 "$ref":  "#/components/schemas/ReportsImportResult"
                                                                                                                                             }
                                                                                                                              }
                                                                                                     }
                                                                                     },
                                                                             "422":  {
                                                                                         "description":  "Validation Error",
                                                                                         "content":  {
                                                                                                         "application/json":  {
                                                                                                                                  "schema":  {
                                                                                                                                                 "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                             }
                                                                                                                              }
                                                                                                     }
                                                                                     }
                                                                         }
                                                       }
                                          },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "AnalyzeRequest":  {
                                                              "properties":  {
                                                                                 "resume_text":  {
                                                                                                     "type":  "string",
                                                                                                     "minLength":  20,
                                                                                                     "title":  "Resume Text"
                                                                                                 },
                                                                                 "jd_text":  {
                                                                                                 "type":  "string",
                                                                                                 "minLength":  20,
                                                                                                 "title":  "Jd Text"
                                                                                             }
                                                                             },
                                                              "type":  "object",
                                                              "required":  [
                                                                               "resume_text",
                                                                               "jd_text"
                                                                           ],
                                                              "title":  "AnalyzeRequest"
                                                          },
...
                                       "HTTPValidationError":  {
                                                                   "properties":  {
                                                                                      "detail":  {
                                                                                                     "items":  {
                                                                                                                   "$ref":  "#/components/schemas/ValidationError"
                                                                                                               },
                                                                                                     "type":  "array",
                                                                                                     "title":  "Detail"
                                                                                                 }
                                                                                  },
                                                                   "type":  "object",
                                                                   "title":  "HTTPValidationError"
                                                               },
...
                                   }
                   }
}